# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Food Delivery API
  version: 1.0.0
  description: REST API for a food delivery service.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/categories:
    get:
      tags:
        - Categories
      summary: List categories
      operationId: listCategories
      security: []
      responses:
        "200":
          description: Category list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: internal server error

    post:
      tags:
        - Categories
      summary: Create a category
      description: Creates a new category. Requires the admin role.
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
            example:
              name: Pizza
              slug: pizza
      responses:
        "201":
          description: Category successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid request body or category data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidBody:
                  value:
                    error: invalid request body
                invalidData:
                  value:
                    error: invalid category data
        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: internal server error
  /api/v1/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"

    get:
      tags:
        - Categories
      summary: Get category by ID
      operationId: getCategoryById
      security: []
      responses:
        "200":
          description: Category found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Invalid category ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid category id
        "404":
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: category not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: internal server error

    put:
      tags:
        - Categories
      summary: Update a category
      description: Replaces category name and slug. Requires the admin role.
      operationId: updateCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
            example:
              name: Italian Pizza
              slug: italian-pizza
      responses:
        "204":
          description: Category successfully updated.
        "400":
          description: Invalid category ID, request body, or category data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: category not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Categories
      summary: Delete a category
      description: Deletes a category. Requires the admin role.
      operationId: deleteCategory
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Category successfully deleted.
        "400":
          description: Invalid category ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid category id
        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Category not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: category not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/products:
    get:
      tags:
        - Products
      summary: List products
      operationId: listProducts
      security: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Values less than or equal to zero are replaced with the default limit.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Product list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "400":
          description: Invalid filter parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - Products
      summary: Create a product
      description: Creates a product. Requires the admin role.
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request body or product data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/products/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"

    get:
      tags:
        - Products
      summary: Get product by ID
      operationId: getProductById
      security: []
      responses:
        "200":
          description: Product found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: product not found
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - Products
      summary: Update a product
      description: Replaces editable product data. Requires the admin role.
      operationId: updateProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid product ID, request body, or product data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Products
      summary: Deactivate a product
      description: Soft-deactivates a product instead of physically deleting it. Requires the admin role.
      operationId: deactivateProduct
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Product successfully deactivated.
        "400":
          description: Invalid product ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/addresses:
    post:
      tags:
        - Addresses
      summary: Create an address
      description: Creates a saved delivery address for the authenticated user.
      operationId: createAddress
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressRequest"
            example:
              label: Home
              city: Amsterdam
              street: Damrak
              house_number: "10"
              apartment_number: "42"
              entrance: "2"
              floor: "5"
              comment: Call before delivery

      responses:
        "201":
          description: Address successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

        "400":
          description: Invalid request body or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Addresses
      summary: List user addresses
      description: Returns saved addresses belonging to the authenticated user.
      operationId: listUserAddresses
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Address list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"

        "400":
          description: Invalid address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/addresses/{id}:
    parameters:
      - $ref: "#/components/parameters/AddressId"

    get:
      tags:
        - Addresses
      summary: Get address by ID
      description: Returns an address belonging to the authenticated user.
      operationId: getAddressById
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Address found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

        "400":
          description: Invalid address ID or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Address not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: address not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags:
        - Addresses
      summary: Update an address
      description: Partially updates an address belonging to the authenticated user.
      operationId: updateAddress
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressRequest"
            example:
              label: Office
              comment: Leave at reception

      responses:
        "200":
          description: Address successfully updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

        "400":
          description: Invalid address ID, request body, or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Address not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: address not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - Addresses
      summary: Delete an address
      description: Deletes an address belonging to the authenticated user.
      operationId: deleteAddress
      security:
        - BearerAuth: []

      responses:
        "204":
          description: Address successfully deleted.

        "400":
          description: Invalid address ID or address data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Address not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: address not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/orders:
    post:
      tags:
        - Orders
      summary: Create an order
      description: Creates an order for the authenticated user using one of their saved delivery addresses.
      operationId: createOrder
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            example:
              address_id: 1
              items:
                - product_id: 10
                  quantity: 2
                - product_id: 15
                  quantity: 1

      responses:
        "201":
          description: Order successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "400":
          description: Invalid request body, address, product, or order data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: unauthorized

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - Orders
      summary: List user orders
      description: Returns orders belonging to the authenticated user.
      operationId: listUserOrders
      security:
        - BearerAuth: []

      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Values less than or equal to zero are replaced with the default limit.

        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

      responses:
        "200":
          description: User order list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

        "400":
          description: Invalid pagination parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/orders/{id}:
    parameters:
      - $ref: "#/components/parameters/OrderId"

    get:
      tags:
        - Orders
      summary: Get user order by ID
      description: Returns an order only if it belongs to the authenticated user.
      operationId: getUserOrderById
      security:
        - BearerAuth: []

      responses:
        "200":
          description: Order found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

        "400":
          description: Invalid order ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: invalid order id

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "404":
          description: Order not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: order not found

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/admin/orders:
    get:
      tags:
        - Admin
      summary: List all orders
      description: Returns a filtered and paginated order list. Requires the admin role.
      operationId: adminListOrders
      security:
        - BearerAuth: []

      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"

        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1

        - name: from
          in: query
          required: false
          description: Start date in YYYY-MM-DD format using the configured application timezone.
          schema:
            type: string
            format: date
          example: "2026-09-01"

        - name: to
          in: query
          required: false
          description: End date in YYYY-MM-DD format using the configured application timezone. The full day is included.
          schema:
            type: string
            format: date
          example: "2026-09-30"

        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

      responses:
        "200":
          description: Filtered order list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"

        "400":
          description: Invalid filter or pagination parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "403":
          description: Admin role is required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CategoryId:
      name: id
      in: path
      required: true
      description: Category identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
    ProductId:
      name: id
      in: path
      required: true
      description: Product identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
    AddressId:
      name: id
      in: path
      required: true
      description: Address identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
    OrderId:
      name: id
      in: path
      required: true
      description: Order identifier.
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    Category:
      type: object
      required:
        - id
        - name
        - slug
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Pizza
        slug:
          type: string
          example: pizza
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCategoryRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          description: Category name. Leading and trailing whitespace is ignored.
          example: Pizza
        slug:
          type: string
          minLength: 1
          description: Category slug. Leading and trailing whitespace is ignored.
          example: pizza
    UpdateCategoryRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          example: Italian Pizza
        slug:
          type: string
          minLength: 1
          example: italian-pizza
...